sync/requirements.txt
flask== 3.1.0
flask-restx==1.3.0
# torch 
transformers==4.48.0 
SentencePiece==0.2.0 
pymongo==3.0.1 
sacremoses==0.1.1 
PyYAML==5.4.1 
protobuf==5.29.3